Remove bad constraint from probes table on chip packages
This change only affects packages for rodents and humans.  It drops
the not null constraint from the accessions collumn which is
problematic since lousy platforms will sometimes have probes that are
measuring "who-knows what".  Users ought to have a right to know when
they are using a probe like this...



git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/AnnotationDbi@27202 bc3139a8-67e5-0310-9ffc-ced21a209358