Adds support to make most chip and organism sqlite files
This update adds many new functions to add functionality to make many
new kinds of organism and chip based sqlite files.  In general there
is a master function for each schema that adds the ability to make
sqlite files of that type (with correct parameters).  These functions
will later be wrapped in another function that should call these
functions and then also call makeAnnDbPkg() to make the sqlite file
and then wrap it up in a new package.




git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/AnnotationDbi@29660 bc3139a8-67e5-0310-9ffc-ced21a209358