Fixed rare bug that can happen when an ID is a very long integer
Fixed a rare bug that can happen where a very long integer is the ID.
When this happens R can promote this to a double before passing it to
SQLite which will then add an unwanted .0 extension onto the end of
it.  Now Sqlforge will check for such IDs and cast them (but only if
needed).


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/AnnotationDbi@30784 bc3139a8-67e5-0310-9ffc-ced21a209358