Updates man pages for org packages, and changes ACCNUM for arabidopsis.
The update to the man pages just attempts to further clarify their
purpose, but the change to all arabidopsis packages, is more
significant.  The arabidopsis change combines the MULTIHIT and ACCNUM
mappings into a single mapping (ACCNUM).  The new mapping should
contain all the information that was previously spread across the two
mappings before.


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/AnnotationDbi@34253 bc3139a8-67e5-0310-9ffc-ced21a209358