Additions to support the creation of safe GO2ALL mappings.
These additions are annotation support for the use of custom GSEABase
GeneSetCollection objects in Category analysis.  Their primary purpose
is to allow that user to safely make sure that a GO to gene mapping is
a GO to ALL style mapping since this is the kind of mapping you need
if you want to do Category analysis with a custome GeneSetCollection
object.  Also these objects allow you to cast a GO to gene style
mapping up to become a GO to ALL style mapping.  We all love
convenience.



git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/AnnotationDbi@41598 bc3139a8-67e5-0310-9ffc-ced21a209358