Bug fix for the GO and GOALL issue
If users requested both GO and GOALL before, then the results would be
messed up for part of the GOALL results.  This fixes that by changing
the orgDb and chipDb select methods so that they use a query that
returns fully qualified field names.  IOW internally it's now more
like this: "SELECT go.go_id AS 'go.go_id'" etc. which results in the
ability to distiguish results that come back when the field names are
identical but the table names are the same.


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/AnnotationDbi@103220 bc3139a8-67e5-0310-9ffc-ced21a209358